Update project structure
