+++ src/components/Content/ContentItem.vue
