+++ src/components/Content/DetailCard.vue
