+++ src/components/Etcetera/SubtitleToCSV.vue
@@ -84,7 +84,7 @@
 </template>
 
 <script>
-import { subtitleAPI } from "../api";
+import { subtitleAPI } from "../../api";
 export default {
   data() {
     return {
